FindAdv: Amend d09f5ce1: Added new debug level :findverbose
Fix the GUI handling for the new value using now 'unsigned long long' instead of 'int'

+++ src/support/debug.cpp
@@ -129,7 +129,7 @@ Debug::Type Debug::value(string const & val)
 			break;
 		// Is it a number?
 		if (isStrInt(tmp))
-			l |= static_cast<Type>(convert<int>(tmp));
+			l |= static_cast<Type>(convert<unsigned long long>(tmp));
 		else
 			// Search for an explicit name
 			for (DebugErrorItem const & item : errorTags)
